+++ src/main/kotlin/dev/aaronhowser/mods/geneticsresequenced/block/machine/incubator/IncubatorBlockEntity.kt
@@ -5,6 +5,7 @@ import dev.aaronhowser.mods.geneticsresequenced.block.base.handler.WrappedHandle
 import dev.aaronhowser.mods.geneticsresequenced.config.ServerConfig
 import dev.aaronhowser.mods.geneticsresequenced.recipe.base.AbstractIncubatorRecipe
 import dev.aaronhowser.mods.geneticsresequenced.recipe.base.IncubatorRecipeInput
+import dev.aaronhowser.mods.geneticsresequenced.recipe.incubator.DupeCellRecipe
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModBlockEntities
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModBlocks
 import dev.aaronhowser.mods.geneticsresequenced.registry.ModItems
@@ -153,6 +154,8 @@ class IncubatorBlockEntity(
             RIGHT_BOTTLE_SLOT_INDEX
         )
 
+        var onlyDupeCellRecipes = true
+
         for (slotIndex in bottleSlots) {
             val bottomStack = itemHandler.getStackInSlot(slotIndex)
 
@@ -160,6 +163,8 @@ class IncubatorBlockEntity(
             val incubatorRecipe = AbstractIncubatorRecipe.getIncubatorRecipe(level!!, recipeInput)
 
             if (incubatorRecipe != null) {
+                if (incubatorRecipe !is DupeCellRecipe) onlyDupeCellRecipes = false
+
                 val output = incubatorRecipe.assemble(recipeInput, level!!.registryAccess())
 
                 if (!output.isEmpty) {
@@ -170,6 +175,8 @@ class IncubatorBlockEntity(
                 val hasMix = potionBrewing.hasMix(bottomStack, topStack)
 
                 if (hasMix) {
+                    onlyDupeCellRecipes = false
+
                     val output = potionBrewing.mix(topStack, bottomStack)
 
                     if (!output.isEmpty) {
@@ -179,7 +186,7 @@ class IncubatorBlockEntity(
             }
         }
 
-        topStack.shrink(1)
+        if (!onlyDupeCellRecipes) topStack.shrink(1)
         setChanged()
     }