+++ CHANGELOG.md
@@ -306,6 +306,7 @@
 
 # 1.5.4
 
+- Incubators now work as Brewing Stands again
 - Added these tags to all custom damage types: `#minecraft:no_impact`, `#minecraft:no_anger`, `#minecraft:no_knockback`
 - Fixed Genes desyncing on client when you change dimensions or respawn (#51)
 - Fixed Set/Unset Anti-Plasmid recipes not working (#50)