+++ CHANGELOG.md
@@ -313,6 +313,6 @@
 - Fixed Set/Unset Anti-Plasmid recipes not working (#50)
 - Fixed Virus recipes not working (#48)
 - Added these tags to all custom damage types: `#minecraft:no_impact`, `#minecraft:no_anger`, `#minecraft:no_knockback`
-- Mod commands can now use any of the following: "/geneticsresequenced", "/genetics", "gr"
-- Renamed the command "/gr clearBioluminescenceBlocks" to "/gr clearBioGlow"
+- Mod commands can now use any of the following: `/geneticsresequenced`, `/genetics`, `/gr`
+- Renamed the command `/gr clearBioluminescenceBlocks` to `/gr clearBioGlow`
 - Updated the Patchouli book to show changes to Flight's default required Genes (#49, #46 kinda)